Extracted EquipmentRepository
